reorg + ajout @valid et @controlleradvice

+++ src/main/java/com/safetynet/alerts/service/PersonServiceImpl.java
@@ -1,8 +1,10 @@
 package com.safetynet.alerts.service;
 
+import com.safetynet.alerts.exception.AlreadyExistsException;
+import com.safetynet.alerts.exception.NoContainException;
+import com.safetynet.alerts.exception.NotFoundException;
 import com.safetynet.alerts.model.Person;
 import com.safetynet.alerts.repository.IPersonRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -10,25 +12,52 @@ import java.util.Optional;
 
 @Service
 public class PersonServiceImpl implements IPersonService {
-    @Autowired
-    private IPersonRepository personRepository;
+    private final IPersonRepository personRepository;
 
-    public Optional<Person> getPerson(String firstname, String lastname) throws IOException {
-        //return Optional.empty();
-        return personRepository.findByFirstNameAndLastName(firstname, lastname);
+    public PersonServiceImpl(IPersonRepository personRepository) {
+        this.personRepository = personRepository;
+    }
+
+    public Person getPerson(String firstname, String lastname) throws NotFoundException, NoContainException, IOException {
+        Optional<Person> person = personRepository.findByFirstNameAndLastName(firstname, lastname);
+        if (person.isPresent()) {
+            return person.get();
+        } else {
+           //throw new NotFoundException("person not found");
+            throw new NoContainException();
+        }
     }
 
     @Override
-    public Person addPerson(Person personToAdd) throws IOException {
-        return personRepository.add(personToAdd);
+    public Person addPerson(Person personToAdd) throws IOException, AlreadyExistsException {
+        Optional<Person> person = personRepository.findByFirstNameAndLastName(personToAdd.getFirstName(), personToAdd.getLastName());
+        if (person.isPresent()) {
+            throw new AlreadyExistsException("person with these firstname and lastname already exist !");
+        } else {
+            return personRepository.add(personToAdd);
+        }
+
     }
     @Override
-    public Person updatePerson(Person currentPerson, Person personToUpdate) throws IOException {
-        return personRepository.update(currentPerson, personToUpdate);
+    public Person updatePerson(String firstname, String lastname, Person personToUpdate) throws IOException, NotFoundException {
+        Optional<Person> person = personRepository.findByFirstNameAndLastName(firstname, lastname);
+        if (person.isPresent()) {
+            Person currentPerson = person.get();
+            return personRepository.update(currentPerson, personToUpdate);
+        } else {
+            throw new NotFoundException("person not found");
+        }
     }
 
     @Override
-    public void deletePerson(Person personToDelete) throws IOException {
-        personRepository.delete(personToDelete);
+    public void deletePerson(String firstname, String lastname) throws IOException, NotFoundException {
+
+        Optional<Person> person = personRepository.findByFirstNameAndLastName(firstname, lastname);
+        if (person.isPresent()) {
+            Person personToDelete = person.get();
+            personRepository.delete(personToDelete);
+        } else {
+            throw new NotFoundException("person not found");
+        }
     }
 }
